chore: refactor parse.cloud.define
