+++ _tests_/integration/crud.test.js
@@ -15,7 +15,7 @@ test('Post Exists', async () => {
 
 test('Post object - async', async () => {
   // expect.assertions(1);
-  const post_params = {
+  const postParams = {
     parseClass: 'SurveyData',
     signature: 'Joe',
     photoFile: 'pictureofJoe',
@@ -24,7 +24,7 @@ test('Post object - async', async () => {
       lname: 'Bouble',
     },
   };
-  return cloudFunctions.postObjectsToClass(post_params).then((result) => {
+  return cloudFunctions.postObjectsToClass(postParams).then((result) => {
     console.log(result);
   });
-});
+}, 30000);
